Add hidden attribute to grants
Currently the only place the "hidden" attribute is used is in the voter
signup page.  The intent is that some grants may be private and only voted
on by the Art Core.  Admins should modify individual voters to add them
to hidden grants.

Also discovered the ruby ".pluck" function.  Super useful and not at all
obvious!  Fixes #40